+++ CHANGELOG.md
@@ -14,6 +14,8 @@ Fixed
 - Issue causing site cells to get unordered when deleting and resetting in the same editting session.
 - Issue where re-ordering sites would unnecessarily change the next site index.
 - Issue where site image would not up-date when changing site names to known sites.
+- Issue where controls became disabled if opening the Site picker in the Hormone Detail view, navigating away,
+  and then returning.
 
 # 3.7.0